create partitions tables
