+++ test-docker-compose.yml
@@ -11,7 +11,7 @@ services:
       - "5432:5432"
 
   postgres_pv:
-    image: postgres:14.2
+    image: postgres:14.5
     restart: always
     environment:
       - POSTGRES_USER=postgres
